Rename the config wizard: bartleby ready -> bartleby config
The interactive settings wizard was `bartleby ready`, but "ready" better
describes getting the skill installed into the harness. Rename the command
(and commands/ready.py -> commands/config.py, tests/test_ready.py ->
tests/test_config.py) to free up `ready` for the skill installer landing
next. Hard rename, no alias, per the no-backwards-compat working agreement.
